pyproject.toml
[tool.poetry]
name = "discord-tft"
version = "0.1.0"
description = "Discord bot to get tft stats and comps"
authors = ["Ryan Bui <ryan.bui@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-utils-traveler = "^0.0.5"
discord = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
